refactor(ui): remove dead components from Storybook
Remove 8 unused components (MatchCountdown, MatchReport, MatchResult,
MatchStats, MatchesOverview, MatchesTabs, CoachProfile, UpcomingMatches)
and the SponsorGrid story. Relocate shared match mock data to
components/match/match.mocks.ts.

Closes #979
