--- apps/web/src/components/home/UpcomingMatches/index.ts
@@ -1,5 +1,0 @@
-/**
- * UpcomingMatches Component Exports
- */
-export { UpcomingMatches } from "./UpcomingMatches";
-export type { UpcomingMatch, UpcomingMatchesProps } from "./UpcomingMatches";
