+++ apps/web/src/components/home/index.ts
@@ -14,9 +14,6 @@ export type {
 export { NewsCard } from "@/components/article/NewsCard";
 export type { NewsCardProps } from "@/components/article/NewsCard";
 
-export { UpcomingMatches } from "./UpcomingMatches";
-export type { UpcomingMatchesProps, UpcomingMatch } from "./UpcomingMatches";
-
 export { MatchWidget } from "./MatchWidget";
 export type { MatchWidgetProps } from "./MatchWidget";
