--- apps/web/src/components/match/MatchCountdown/index.ts
@@ -1,1 +1,0 @@
-export { MatchCountdown, type MatchCountdownProps } from "./MatchCountdown";
