+++ apps/web/src/components/match/MatchList/MatchList.stories.tsx
@@ -11,7 +11,7 @@ import {
   mockFinishedMatch,
   mockPostponedMatch,
   mockMatches,
-} from "@/components/home/UpcomingMatches/UpcomingMatches.mocks";
+} from "@/components/match/match.mocks";
 
 const KCVV_ID = 1235;
