--- apps/web/src/components/match/MatchReport/index.ts
@@ -1,5 +1,0 @@
-export {
-  MatchReport,
-  type MatchReportProps,
-  type MatchReportPhoto,
-} from "./MatchReport";
