--- apps/web/src/components/match/MatchResult/index.ts
@@ -1,5 +1,0 @@
-export {
-  MatchResult,
-  type MatchResultProps,
-  type MatchResultTeam,
-} from "./MatchResult";
