--- apps/web/src/components/match/MatchStats/index.ts
@@ -1,6 +1,0 @@
-export {
-  MatchStats,
-  type MatchStatsProps,
-  type MatchStatsData,
-  type StatPair,
-} from "./MatchStats";
