+++ apps/web/src/components/match/index.ts
@@ -11,13 +11,7 @@ export type { MatchDetailViewProps } from "./MatchDetailView";
 export { MatchList } from "./MatchList/MatchList";
 export type { MatchListProps } from "./MatchList/MatchList";
 
-export { MatchesOverview } from "./MatchesOverview/MatchesOverview";
-export type { MatchesOverviewProps } from "./MatchesOverview/MatchesOverview";
-
 export { MatchesSlider } from "./MatchesSlider/MatchesSlider";
 export type { MatchesSliderProps } from "./MatchesSlider/MatchesSlider";
-
-export { MatchesTabs } from "./MatchesTabs/MatchesTabs";
-export type { MatchesTabsProps } from "./MatchesTabs/MatchesTabs";
 
 export type { UpcomingMatch } from "./types";
