--- apps/web/src/components/team/CoachProfile/index.ts
@@ -1,1 +1,0 @@
-export { CoachProfile, type CoachProfileProps } from "./CoachProfile";
